Fixed message parts array potentially out of sync with text view.
Since the array is public we need to make it immutable and readonly.
Also, since it could be accessed at any time, it needs to be generated
when called. But to avoid regenerating it when unnecessary (i.e. when
the content hasn't since changed) we also store the attributed string it
was based on to compare against when the array is next requested.