Changed address table height constraint and inset to bottom constraint.
Since the bottom content inset for the table view was being set to the
negative value of controllerYOffset and that property is a NSUInteger,
the end result was a huge bottom content inset. With that issue
corrected, we then need to handle the keyboard overlapping the table
view. The simplest way (and most consistent with the participant picker)
is to have the keyboard dismiss on dragging.