Moved address height constraint to parent VC and changed to full height.
Instead of adjusting the view's height depending on the display of the
search results table (and allowing the child view controller to make
height calculations based on its parent view controller), we instead let
the view be the full height. We then use a UIView subclass as the
address bar view controller's root view so that we can ignore touches
directly to it as opposed to its subviews. That is, since the view now
takes up the full height, when the table view is hidden, we want touches
to pass through to the parent VC / superview.