Fixed read status displayed for old messages as new messages come in.
Each message's status is displayed as a collection view section footer
and the footer's content depends on other messages. So even when a
message isn't updated, its footer may need to be updated. We could
simply reload the entire collection view, but we want to display changes
through animations. We could reload all the visible sections, but that
causes a needless and distracting cross-fade animation of the entire
section (i.e. message bubble and its footer). Instead, we reconfigure
the footers ourselves just as we do when first setting up a footer.

However, UICollectionView doesn't provide a way to get a footer that is
already onscreen like it does for cells. So we weakly hold on to all
configured footers using a hash table.

Since a section's index can change over time (e.g. through moving or
other sections being inserted or deleted), we can't simply keep track of
the sections by their indexes. Instead, we store the associated message
for each footer using a property of the footer. We can then reconfigure
the footer for that message each time the query controller reports any
changes.