Refactored message composing UI to work around layout/scrolling bugs.
The main changes are: Fixes to many layout and scrolling issues by
switching to manual layout to work around apparent Auto Layout issues
when used in that context, i.e. an auto height resizing text view in an
input accessory view. Using a UILabel to show placeholder text instead
of using the text view's own text. Using the custom placeholder text if
it is provided.