Worked around conversation view controller never deallocated.
It seems using inputAccessoryView on a view controller causes a retain
cycle. To work around this, we instead use the view controller's view's
inputAccessoryView property. This also solves the issue of the
placeholder text and send button not appearing during the navigation
controller's push animation after the first appearance on iOS 7.1 but
the issue remains on iOS 8.

However, this change has drawbacks:

On iOS 8, the toolbar doesn't animate onscreen the first time. But
subsequent conversation view controllers will have their toolbars
animate onscreen.

On iOS 7.1, when starting to swipe from the left edge to go back, the
toolbar does not begin to slide off with the conversation content. That
is, the toolbar remains at the bottom of the screen. If the swipe
completes then the toolbar slides off to the right separately.