Fixed automatic scrolling to bottom when message input is active.
We can't keep track of where the keyboard is onscreen by simply toggling
a boolean on each notification since notification is posted in a number
of different situations, e.g. when the iOS 8 QuickType bar is collapsed
or expanded. Instead, we use a UITextView notification to precisely know
when the text view becomes the first responder.