Fixed conversation content insets for keyboard and typing indicator.
This fixes: Not scrolling to the bottom on the first appearance. Making
room for the keyboard when the keyboard doesn't need room due to an
external keyboard. Leaving room for the keyboard when returning to the
conversation even when the keyboard is not shown.

Also, we now always keep space for the typing indicator (i.e. even when
it is not shown) and we show it even when not scrolled to the bottom
(i.e. so that it is always just above the input toolbar).