Fixed typing began status sent only once until text is cleared.
This is issue was created with 13cc331c because it assumed that the
message toolbar should only notify its delegate when typing begins and
ends. However, LayerKit automatically changes the typing status to
paused and finished after a while if it hasn't since received a began
status. So we need to continually notify the delegate that typing is
occurring. To prevent that confusion going forward, we rename the
delegate method accordingly.