Changed typing indicator content to be dependent on available space.
Our first preference is to display all the full names of participants
who are typing. But if that content doesn't fit then we want to display
all of their first names. If that doesn't fit then we we want to display
as many of their first names as possible with a summary of the names not
shown (e.g. 'and 3 others').