Fixed crash on receiving query controller callback after niling out.
(APPS-853)

The crash is caused by the following sequence:

1. Changes are received by the query controller on a background thread
which asynchronously queues the changes to be delivered to the query
controller on the main thread.
2. The view controller disappears and clears self.queryController.
3. The query controller delivers the changes to view controller.
4. The view controller attempts to get the number of items from
self.queryController in the UICollectionView delegate/data source
methods and returns zero since self.queryController is nil.

If we wanted to keep the behavior of niling out the query controller
when the view disappears, we could just nil out the delegate as well or
check that the passed query controller == self.queryController when
handling query controller delegate callbacks before processing changes.

But instead we want to keep the query controller around when the view
disappears.