Fixed typing indicator shown for wrong conversation. (APPS-961)
We had been trying to limit the typing indicator notifications to the
current conversation by passing the conversation in our call to add a
observer when setting up the view controller. However, if the view
controller is initialized without a conversation then we'd receive
typing indicator notifications for all conversations.

So instead we explicitly pass nil when adding the observer but then when
handling the notification, we ignore notifications from other
conversations. And to handle the case of the conversation later being
changed to a different conversation, we clear out received typing
indicators when the switch occurs.