Fix Conversation View's insets in popover
When the conversation view is in a popover and the keyboard is toggled, the conversation view's insets don't need to be adjusted because the popover moves to make room for the keyboard. I'm not exactly sure if this is the right solution, but this is what we implemented, so I figured I'd share it with you and let you decide.